Correct a bug occurring when the underlying file object is filling very fast.
It is possible that « unused_raw » contains more than one end of stream and start of another. We need to restart the decoder while there are unused data at the end.

This caused a strange bug when using that to stream a commoncrawl archive on my server with a good bandwidth, that don't happened on my test computer :
zlib.error: Error -3 while decompressing: incorrect header check (in restart_decoder).